+++ source/index.js
@@ -1,18 +1,16 @@
-const stringCase = require('case');
-
-const requireContext = require.context(
-  './components',
-  true,
-  /\.\/[^/]+\/index\.js$/,
-);
-
-const startWithCapitalLetter = (string) =>
-  `${string[0].toUpperCase()}${string.slice(1)}`;
-
-requireContext.keys().forEach((key) => {
-  const componentName = startWithCapitalLetter(
-    stringCase.camel(key.replace(/^.+\/([^/]+)\/index\.js/, '$1')),
-  );
-
-  module.exports[componentName] = requireContext(key).default;
-});
+export { default as Badge } from './components/badge';
+export { default as BaseSelect } from './components/base-select';
+export { default as Button } from './components/button';
+export { default as Checkbox } from './components/checkbox';
+export { default as FieldHint } from './components/field-hint';
+export { default as FieldLabel } from './components/field-label';
+export { default as Form } from './components/form';
+export { default as FormRow } from './components/form-row';
+export { default as Hr } from './components/hr';
+export { default as Input } from './components/input';
+export { default as Logo } from './components/logo';
+export { default as Meter } from './components/meter';
+export { default as Page } from './components/page';
+export { default as Panel } from './components/panel';
+export { default as Select } from './components/select';
+export { default as Text } from './components/text';